// Parcial2/src/type/types.ts
export interface AttrChunorris{
    animal: string;
    carrer:string;
    celebrity: string;
    dev: string;
    explicity: string;
    fashion: string;
    history: string;
    money: string;
    movie: string;
    music: string;
    political: string;
    religion: string;
    science: string;
    sport: string;
    travel: string;
}